+++ app/src/main/java/net/kuama/documentscanner/presentation/ScannerActivity.kt
@@ -1,12 +1,10 @@
 package net.kuama.documentscanner.presentation
 
-import android.annotation.SuppressLint
 import android.graphics.Bitmap
 import android.widget.Toast
 import net.kuama.documentscanner.R
 import net.kuama.documentscanner.exceptions.NullCorners
 
-@SuppressLint("UnsafeExperimentalUsageError")
 class ScannerActivity : BaseScannerActivity() {
     override fun onError(throwable: Throwable) {
         when (throwable) {